Updated hello4.py to make it more comprehensive.
There are DEBUG print statements in simpleth that need
to be removed sometime.

I'm still stuck on that get_new_events() test. Starting
to feel like something gets out of sync. If I try to
connect to the test contract from an interpeter window,
the address of the contract is different and there are
a few extra blocks on the chain. Wondering if somewhere
I'm getting the previously deployed contract instead
of the most current contract in my testing (or in
simpleth.)

Doing a commit to catch all the latest edits that are
working.