Renamed nat2rtd.py to nat2rst.py - This is more accurate.
Started to add a param to nat2rst.py for the heading
underline chars. It was too clumsy to try to pass
special chars as an arg. Decided it wasn't worth it.
Stick to the global constants for those chars. Easy
enough to change. Undid all those edits.

Along the way I made some tweaks to documentation.