Add Ganache URL as an environment variable.
Add url as a Blockchain property.

Add test cases to check the property.

Add Blockchain class comments about the environment
variable.

Add a new document, TestContract.rst, that sucks in the
source file. It is now the last document in the ToC.

Minor tweaks along the way.

This does it for all my ideas on the code.

Checkall runs OK.